Refactoring cli
